move copy button to component
